# fronted/OPENAPI/ExamSystemApi.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ExamSystemApi",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/answers/{page}/{size}": {
      "get": {
        "summary": "findAllQuestion",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "查询所有题库",
                      "data": {
                        "records": [
                          {
                            "question": "",
                            "subject": "",
                            "score": "",
                            "section": "",
                            "level": "",
                            "type": ""
                          }
                        ],
                        "total": 0,
                        "size": 0,
                        "current": 0,
                        "ascs": [
                          ""
                        ],
                        "descs": [
                          ""
                        ],
                        "optimizeCountSql": false,
                        "isSearchCount": false,
                        "searchCount": false,
                        "pages": 0,
                        "asc": [
                          ""
                        ],
                        "desc": [
                          ""
                        ]
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/answers/by-question-id-and-type": {
      "get": {
        "summary": "findByQuestionIdAndType",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "questionId",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "records": [
                        {
                          "questionId": "",
                          "question": "",
                          "subject": "",
                          "score": "",
                          "section": "",
                          "level": "",
                          "type": ""
                        }
                      ],
                      "total": 0,
                      "size": 0,
                      "current": 0,
                      "ascs": [
                        ""
                      ],
                      "descs": [
                        ""
                      ],
                      "optimizeCountSql": false,
                      "isSearchCount": false,
                      "searchCount": false,
                      "pages": 0,
                      "asc": [
                        ""
                      ],
                      "desc": [
                        ""
                      ]
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/answers/by-type": {
      "get": {
        "summary": "findAllByType",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "查询成功",
                      "data": [
                        {
                          "questionId": "",
                          "question": "",
                          "subject": "",
                          "score": "",
                          "answer": "",
                          "section": "",
                          "level": "",
                          "type": ""
                        }
                      ]
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/v1/admin/{adminId}": {
      "get": {
        "summary": "findById",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "adminId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "请求成功",
                      "data": {
                        "adminId": 0,
                        "adminName": "",
                        "sex": "",
                        "tel": "",
                        "email": "",
                        "pwd": "",
                        "cardId": "",
                        "role": ""
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "deleteById",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "adminId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "请求成功",
                      "data": null
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "update",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "adminId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "adminId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "adminName",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sex",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tel",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pwd",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cardId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "请求成功",
                      "data": 0
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/admin": {
      "post": {
        "summary": "add",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "adminId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "adminName",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sex",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tel",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pwd",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cardId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "请求成功",
                      "data": 0
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码，表示一种错误类型\n请求成功，状态码为200"
          },
          "message": {
            "type": "string",
            "description": "对错误代码的详细解释"
          },
          "data": {
            "type": "integer",
            "description": "返回的结果包装在value中，value可以是单个对象"
          }
        }
      },
...
    },
    "securitySchemes": {}
  },
  "servers": []
}